fix(auth): address gate-review route-back for #1120
Mechanical (repo-shape line-count cap, 600):
- split test_sso.py (706→539): the Task-05 self-host tests move to
  test_sso_selfhost.py (reusing test_sso's _connection/_FakeDb fixtures).

Substantive (★):
- T2-AUTH-5 intent regression: seedEnabledOrgSsoConnection created an "enabled"
  org SSO connection with NULL OIDC fields, which the new discovery-completeness
  gate correctly flips to enabled=false — so the entry-point discovery the
  fixture backs saw nothing. Seed a structurally-complete PUBLIC client
  (client id + issuer + token_endpoint_auth_method='none', so no secret /
  ciphertext is needed) → the connection is genuinely startable and discovery
  advertises it again.
- Actionable JIT-disabled screen: add sso_jit_disabled to web-auth-errors.ts
  ("Account not provisioned" → ask an admin to enable JIT or invite you), so the
  server's sso_jit_disabled code renders a Proliferate-owned actionable screen —
  making the report's "actionable screen" claim true rather than dropping it.
  Locked with a web-auth-errors.test.ts case.

GitHub-callback convergence deferral accepted per review (canonical
/auth/github/callback already the live contract; legacy path dead — prune
follow-up).

Tests: server test_sso.py + test_sso_selfhost.py = 24 passed (ruff clean);
web-auth-errors.test.ts = 6 passed. Full T2-AUTH-5 needs the tier-2 stack
(Postgres + booted server); the seed fix is validated by inspection + no new
type errors.
